requirements.txt
async-generator==1.10
attrs==22.2.0
autopep8==2.0.2
certifi==2022.12.7
click==8.1.3
exceptiongroup==1.1.1
future==0.18.3
h11==0.14.0
idna==3.4
iso8601==1.1.0
outcome==1.2.0
Pillow==9.2.0
PyBluez @ /phomemo_d30/pybluez
pycodestyle==2.10.0
pyserial==3.5
PySocks==1.7.1
PyYAML==6.0
selenium==4.8.3
serial==0.0.97
sniffio==1.3.0
sortedcontainers==2.4.0
trio==0.22.0
trio-websocket==0.10.2
urllib3==1.26.15
Wand==0.6.10
wsproto==1.2.0
